+++ src/modules/system/migrate/scripts/index.js
@@ -3,3 +3,4 @@
 export { default as BaseMigration } from "./000-base.js";
 export { default as ActiveEffectsMigration } from "./001-activeEffects.js";
 export { default as FoundryV10Migration } from "./002-foundryV10.js";
+export { default as HotfixEightyThreeDotOne } from "./003-hotfix-0.83.1.js";